move device mesh out
